requirements.txt
flask==3.0.3
flask-limiter==3.5.0
flask-talisman==1.1.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
nltk==3.9.1
matplotlib==3.9.0
seaborn==0.13.2
gunicorn==22.0.0
python-dotenv==1.0.1
